fix: lobe hub list connect  in detail and extract use skill connect hook
