+++ src/features/SkillStore/LobeHubList/Item.tsx
@@ -1,23 +1,14 @@
 'use client';
 
-import { getLobehubSkillProviderById } from '@lobechat/const';
 import { ActionIcon, Block, Dropdown, Flexbox, Icon, Image } from '@lobehub/ui';
 import { App } from 'antd';
 import { createStyles, cssVar } from 'antd-style';
 import type { Klavis } from 'klavis';
 import { Loader2, MoreVerticalIcon, Plus, Unplug } from 'lucide-react';
-import React, { memo, useCallback, useEffect, useRef, useState } from 'react';
+import React, { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import { useToolStore } from '@/store/tool';
-import { klavisStoreSelectors, lobehubSkillStoreSelectors } from '@/store/tool/selectors';
-import { KlavisServerStatus } from '@/store/tool/slices/klavisStore';
-import { LobehubSkillStatus } from '@/store/tool/slices/lobehubSkillStore/types';
-import { useUserStore } from '@/store/user';
-import { userProfileSelectors } from '@/store/user/selectors';
-
-const POLL_INTERVAL_MS = 1000;
-const POLL_TIMEOUT_MS = 15_000;
+import { useSkillConnect } from './useSkillConnect';
 
 const useStyles = createStyles(({ css, token }) => ({
   container: css`
@@ -69,12 +60,16 @@ interface ItemProps {
 }
 
 const Item = memo<ItemProps>(
-  ({ description, icon, identifier, isConnected, label, onOpenDetail, serverName, type }) => {
+  ({ description, icon, identifier, label, onOpenDetail, serverName, type }) => {
     const { t } = useTranslation('setting');
     const { styles } = useStyles();
     const { modal } = App.useApp();
-    const [isConnecting, setIsConnecting] = useState(false);
-    const [isWaitingAuth, setIsWaitingAuth] = useState(false);
+
+    const { handleConnect, handleDisconnect, isConnected, isConnecting } = useSkillConnect({
+      identifier,
+      serverName,
+      type,
+    });
 
     // Get localized description
     const i18nPrefix = type === 'klavis' ? 'tools.klavis.servers' : 'tools.lobehubSkill.providers';
@@ -83,215 +78,14 @@ const Item = memo<ItemProps>(
       defaultValue: description,
     });
 
-    const oauthWindowRef = useRef<Window | null>(null);
-    const windowCheckIntervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
-    const pollIntervalRef = useRef<ReturnType<typeof setInterval> | null>(null);
-    const pollTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-
-    // LobeHub skill hooks
-    const checkLobehubStatus = useToolStore((s) => s.checkLobehubSkillStatus);
-    const revokeLobehubConnect = useToolStore((s) => s.revokeLobehubSkill);
-    const getAuthorizeUrl = useToolStore((s) => s.getLobehubSkillAuthorizeUrl);
-    const lobehubServer = useToolStore(
-      lobehubSkillStoreSelectors.getServerByIdentifier(identifier),
-    );
-
-    // Klavis hooks
-    const userId = useUserStore(userProfileSelectors.userId);
-    const createKlavisServer = useToolStore((s) => s.createKlavisServer);
-    const refreshKlavisServerTools = useToolStore((s) => s.refreshKlavisServerTools);
-    const removeKlavisServer = useToolStore((s) => s.removeKlavisServer);
-    const klavisServer = useToolStore(klavisStoreSelectors.getServerByIdentifier(identifier));
-
-    const cleanup = useCallback(() => {
-      if (windowCheckIntervalRef.current) {
-        clearInterval(windowCheckIntervalRef.current);
-        windowCheckIntervalRef.current = null;
-      }
-      if (pollIntervalRef.current) {
-        clearInterval(pollIntervalRef.current);
-        pollIntervalRef.current = null;
-      }
-      if (pollTimeoutRef.current) {
-        clearTimeout(pollTimeoutRef.current);
-        pollTimeoutRef.current = null;
-      }
-      oauthWindowRef.current = null;
-      setIsWaitingAuth(false);
-    }, []);
-
-    useEffect(() => {
-      return () => {
-        cleanup();
-      };
-    }, [cleanup]);
-
-    useEffect(() => {
-      const connected =
-        type === 'lobehub'
-          ? lobehubServer?.status === LobehubSkillStatus.CONNECTED
-          : klavisServer?.status === KlavisServerStatus.CONNECTED;
-
-      if (connected && isWaitingAuth) {
-        cleanup();
-      }
-    }, [type, lobehubServer?.status, klavisServer?.status, isWaitingAuth, cleanup]);
-
-    // Listen for OAuth success message from popup window (for LobeHub skills)
-    useEffect(() => {
-      if (type !== 'lobehub') return;
-
-      const handleMessage = async (event: MessageEvent) => {
-        if (event.origin !== window.location.origin) return;
-
-        if (
-          event.data?.type === 'LOBEHUB_SKILL_AUTH_SUCCESS' &&
-          event.data?.provider === identifier
-        ) {
-          cleanup();
-          await checkLobehubStatus(identifier);
-        }
-      };
-
-      window.addEventListener('message', handleMessage);
-      return () => window.removeEventListener('message', handleMessage);
-    }, [type, identifier, cleanup, checkLobehubStatus]);
-
-    const startFallbackPolling = useCallback(
-      (serverIdOrName: string) => {
-        if (pollIntervalRef.current) return;
-
-        pollIntervalRef.current = setInterval(async () => {
-          try {
-            if (type === 'lobehub') {
-              await checkLobehubStatus(serverIdOrName);
-            } else {
-              await refreshKlavisServerTools(serverIdOrName);
-            }
-          } catch (error) {
-            console.error('[SkillStore] Failed to check status:', error);
-          }
-        }, POLL_INTERVAL_MS);
-
-        pollTimeoutRef.current = setTimeout(() => {
-          if (pollIntervalRef.current) {
-            clearInterval(pollIntervalRef.current);
-            pollIntervalRef.current = null;
-          }
-          setIsWaitingAuth(false);
-        }, POLL_TIMEOUT_MS);
-      },
-      [type, checkLobehubStatus, refreshKlavisServerTools],
-    );
-
-    const startWindowMonitor = useCallback(
-      (oauthWindow: Window, serverIdOrName: string) => {
-        windowCheckIntervalRef.current = setInterval(async () => {
-          try {
-            if (oauthWindow.closed) {
-              if (windowCheckIntervalRef.current) {
-                clearInterval(windowCheckIntervalRef.current);
-                windowCheckIntervalRef.current = null;
-              }
-              oauthWindowRef.current = null;
-              // Check status and then reset waiting state
-              if (type === 'lobehub') {
-                await checkLobehubStatus(serverIdOrName);
-              } else {
-                await refreshKlavisServerTools(serverIdOrName);
-              }
-              setIsWaitingAuth(false);
-            }
-          } catch {
-            if (windowCheckIntervalRef.current) {
-              clearInterval(windowCheckIntervalRef.current);
-              windowCheckIntervalRef.current = null;
-            }
-            startFallbackPolling(serverIdOrName);
-          }
-        }, 500);
-      },
-      [type, checkLobehubStatus, refreshKlavisServerTools, startFallbackPolling],
-    );
-
-    const openOAuthWindow = useCallback(
-      (oauthUrl: string, serverIdOrName: string) => {
-        cleanup();
-        setIsWaitingAuth(true);
-
-        const oauthWindow = window.open(oauthUrl, '_blank', 'width=600,height=700');
-        if (oauthWindow) {
-          oauthWindowRef.current = oauthWindow;
-          startWindowMonitor(oauthWindow, serverIdOrName);
-        } else {
-          startFallbackPolling(serverIdOrName);
-        }
-      },
-      [cleanup, startWindowMonitor, startFallbackPolling],
-    );
-
-    // Handle connect for LobeHub
-    const handleLobehubConnect = async () => {
-      if (lobehubServer?.isConnected) return;
-
-      setIsConnecting(true);
-      try {
-        const provider = getLobehubSkillProviderById(identifier);
-        if (!provider) return;
-
-        const redirectUri = `${window.location.origin}/oauth/callback/success?provider=${encodeURIComponent(identifier)}`;
-        const { authorizeUrl } = await getAuthorizeUrl(identifier, { redirectUri });
-        openOAuthWindow(authorizeUrl, identifier);
-      } catch (error) {
-        console.error('[SkillStore] Failed to get authorize URL:', error);
-      } finally {
-        setIsConnecting(false);
-      }
-    };
-
-    // Handle connect for Klavis
-    const handleKlavisConnect = async () => {
-      if (!userId || !serverName) return;
-      if (klavisServer) return;
-
-      setIsConnecting(true);
-      try {
-        const newServer = await createKlavisServer({
-          identifier,
-          serverName,
-          userId,
-        });
-
-        if (newServer) {
-          if (newServer.isAuthenticated) {
-            await refreshKlavisServerTools(newServer.identifier);
-          } else if (newServer.oauthUrl) {
-            openOAuthWindow(newServer.oauthUrl, newServer.identifier);
-          }
-        }
-      } catch (error) {
-        console.error('[SkillStore] Failed to connect server:', error);
-      } finally {
-        setIsConnecting(false);
-      }
-    };
-
-    const handleConnect = type === 'lobehub' ? handleLobehubConnect : handleKlavisConnect;
-
-    const handleDisconnect = () => {
+    const confirmDisconnect = () => {
       modal.confirm({
         cancelText: t('cancel', { ns: 'common' }),
         centered: true,
         content: t('tools.lobehubSkill.disconnectConfirm.desc', { name: label }),
         okButtonProps: { danger: true },
         okText: t('tools.lobehubSkill.disconnect'),
-        onOk: async () => {
-          if (type === 'lobehub' && lobehubServer) {
-            await revokeLobehubConnect(lobehubServer.identifier);
-          } else if (type === 'klavis' && klavisServer) {
-            await removeKlavisServer(klavisServer.identifier);
-          }
-        },
+        onOk: handleDisconnect,
         title: t('tools.lobehubSkill.disconnectConfirm.title', { name: label }),
       });
     };
@@ -304,7 +98,7 @@ const Item = memo<ItemProps>(
     };
 
     const renderAction = () => {
-      if (isConnecting || isWaitingAuth) {
+      if (isConnecting) {
         return <ActionIcon icon={Loader2} loading />;
       }
 
@@ -317,7 +111,7 @@ const Item = memo<ItemProps>(
                   icon: <Icon icon={Unplug} />,
                   key: 'disconnect',
                   label: t('tools.lobehubSkill.disconnect'),
-                  onClick: handleDisconnect,
+                  onClick: confirmDisconnect,
                 },
               ],
             }}
